Register NEW TVL product under DANZO project
Greetings friends! Regarding our project: https://defillama.com/protocol/danzo-arena?staking=true

That was integrated via this PR https://github.com/DefiLlama/DefiLlama-Adapters/pull/10893

We would like to make some changes to the image etc BUT also register a brand new product which increases the project TVL

So the first product is the arena and the second is a casino, what should be our approach to have both products under the same protocol which we call the Danzo APP (app includes arena and casino). Also could our TVL show without enabling staking as it is inaccurate?

The TVL is not by any means *staked*. It's locked within an address!