Create OpenDeltaOPB.js adapter
Adding the adapter for OpenDelta Perpetual Bond tokens (OPB) on Solana. 

To call the Token2022 library function `amountToUIAmount` we need a signing-account even though it is just a reading on-chain function call. DeFi Llama support explained to us we can use a server side signing-account which the DeFi Llama team can add.