Add Astake to project.
Astake is built on a secure, reliable framework, making it simple for users to take part in Astar dApp Staking. Our protocol is designed enhance the overall liquidity and stability of the ecosystem. By leveraging advanced security measures and a user-friendly interface, Astake aims to become an integral part of Astar's and Soneium’s financial landscape.